moved the model into lib/db because that makes more sense
