+++ api/handler_get_new_packages.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/gophr-pm/gophr/lib/db"
 	"github.com/gophr-pm/gophr/lib/errors"
-	"github.com/gophr-pm/gophr/lib/model/package"
+	"github.com/gophr-pm/gophr/lib/db/model/package"
 )
 
 const (
