+++ lib/db/model/package/assert.go
