+++ lib/db/model/package/is_awesome.go
