+++ router/package_request.go
@@ -8,10 +8,10 @@ import (
 	"github.com/gophr-pm/gophr/lib"
 	"github.com/gophr-pm/gophr/lib/config"
 	"github.com/gophr-pm/gophr/lib/db"
+	"github.com/gophr-pm/gophr/lib/db/model/package/archive"
 	"github.com/gophr-pm/gophr/lib/depot"
 	"github.com/gophr-pm/gophr/lib/github"
 	"github.com/gophr-pm/gophr/lib/io"
-	"github.com/gophr-pm/gophr/lib/model/package/archive"
 	"github.com/gophr-pm/gophr/lib/verdeps"
 )
